문의 게시판, 예약 dao_dto 업로드

+++ SemiTeamProject/WebContent/question/question_content.jsp
@@ -38,6 +38,7 @@
 		    }
 	    }
     	
+	    boolean isUser = user.getMember_no() == qvdto.getQue_writer();
     	
     %>
 <jsp:include page="/template/nav.jsp"></jsp:include>
@@ -116,14 +117,16 @@
 	<div class="row-empty"></div>
 
 	<div class="row right">
-		<% if (sdto == null) { %>
+		<% if (sdto == null && isUser) { %>
 			<a href="question_edit.jsp?que_no=<%= que_no %>">
 				<input class="form-btn form-inline" type="button" value="문의수정">
 			</a>
 		<%} %>
-		<a href="question_delete.do?que_pension_no=<%= que_pension_no %>&que_no=<%= que_no %>">
-			<input class="form-btn form-inline" type="button" value="삭제">
-		</a>
+		<% if(isUser) { %>
+			<a href="question_delete.do?que_pension_no=<%= que_pension_no %>&que_no=<%= que_no %>">
+				<input class="form-btn form-inline" type="button" value="삭제">
+			</a>
+		<%} %>
 		<a href="question_write.jsp">
 			<input class="form-btn form-inline" type="button" value="글쓰기">
 		</a>
